+++ test/parser/basic.js
@@ -1,16 +1,6 @@
 'use strict';
 
-var Inquisitor = require('../src/index');
-
-describe('Initialization', function () {
-  it('should throw an error if no keywords are given and not all keywords are allowed', function () {
-    (function () {
-      Inquisitor.createParser({
-        allowAllKeys: false
-      });
-    }).should.throw();
-  });
-});
+var Inquisitor = require('../../src/index');
 
 describe('Basic parsing', function () {
   beforeEach(function () {
@@ -46,11 +36,12 @@ describe('Basic parsing', function () {
   });
 
   it('should handle non-alphabetical characters correctly', function () {
-    var input = 'foo:/banana';
+    var input = 'foo:/banana bar:5';
     var output = this.parser.parse(input);
 
     output.should.deep.equal({
-      foo: '/banana'
+      foo: '/banana',
+      bar: '5'
     });
   });
 
@@ -69,30 +60,6 @@ describe('Basic parsing', function () {
 
     output.should.deep.equal({
       foo: 'quoted string'
-    });
-  });
-});
-
-describe('Keyword-specific options', function () {
-  describe('multiple', function () {
-    beforeEach(function () {
-      this.parser = Inquisitor.createParser({
-        pairs: [
-          {
-            key: 'foo',
-            multipleValues: true
-          }
-        ]
-      });
-    });
-
-    it('should add multiple values to a key if the option is set to true', function () {
-      var input = 'foo:one bar:baz foo:two';
-      var output = this.parser.parse(input);
-
-      output.should.deep.equal({
-        foo: ['one', 'two']
-      });
     });
   });
 });
